% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LaserIncidents.R
\docType{package}
\name{LaserIncidents}
\alias{LaserIncidents}
\alias{LaserIncidents-package}
\title{Laser Incidents Reported by the FAA}
\format{
A dataframe with 33296 elements
\describe{
  \item{aircraft}{Airplane Name}
  \item{airport}{Airport Location}
  \item{altitude}{Airplane Height (in ft)}
  \item{city}{Airport City}
  \item{flight_id}{Flight ID}
  \item{incident_date}{Date of Laser Pointer Incident (Y/M/D)}
  \item{injury}{Injuries noted due to Laser}
  \item{laser_color}{Laser Color}
  \item{state}{Airport State}
}
}
\source{
\url{https://www.faa.gov/about/initiatives/lasers/laws/}
}
\usage{
LaserIncidents
}
\description{
This package contains specifications for Laser incidents
reported by Federal Aviation Administration

A data set containing Laser Incidents reported by
the Federal Aviation Administration
from 2015-2019
}
\keyword{datasets}
